+++ react/typings/css.d.ts
@@ -1,0 +1,7 @@
+declare module '*.css' {
+  const styles: {
+    [key: string]: string
+  }
+
+  export default styles
+}