+++ react/typings/vtex.styleguide.d.ts
@@ -1,0 +1,5 @@
+declare module 'vtex.styleguide' {
+  export const ButtonWithIcon
+  export const IconClear
+  export const IconPlusLines
+}