+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Buy Together Shelf component.
+
 ## [0.1.0] - 2020-07-23
 
 ### Added